% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brava_trait_path.R
\name{brava_trait_path}
\alias{brava_trait_path}
\title{Get Path to BRAVA Phenotypes File}
\usage{
brava_trait_path()
}
\value{
A character string with the full path to the 'brava_phenotypes.txt' file.
}
\description{
This function returns the file path to the 'brava_phenotypes.txt' file included in the 'bravastring' package.
}
\examples{
brava_file_path <- brava_trait_path()
print(brava_file_path)
}
