% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inflation.R
\name{calc_inflation}
\alias{calc_inflation}
\title{Calculate Inflation Factor}
\usage{
calc_inflation(P, cutoff = 0.5)
}
\arguments{
\item{P}{A numeric vector of p-values.}

\item{cutoff}{A numeric value for the cutoff to use in the chi-squared distribution. Default is 0.5.}
}
\value{
A numeric value representing the inflation factor.
}
\description{
This function calculates the inflation factor for a given set of p-values.
}
\examples{
p_values <- c(0.05, 0.01, 0.03, 0.2)
calc_inflation(p_values)
}
