% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights_meta.R
\name{calc_weights_meta}
\alias{calc_weights_meta}
\title{Calculate Weights for Meta-Analysis}
\usage{
calc_weights_meta(dt, is_inv_var = TRUE, n_eff_name = NULL, se_name = NULL)
}
\arguments{
\item{dt}{A data table or data frame.}

\item{is_inv_var}{Logical. If TRUE, use inverse variance weighting; if FALSE, use square root of effective sample size.}

\item{n_eff_name}{Character. Column name for effective sample size. Used when is_inv_var is FALSE.}

\item{se_name}{Character. Column name for standard errors. Used when is_inv_var is TRUE.}
}
\value{
Data table with added 'weights' column.
}
\description{
Calculate Weights for Meta-Analysis
}
