% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_path.R
\name{canonical_path}
\alias{canonical_path}
\title{Get Path to Canonical Transcripts File}
\usage{
canonical_path()
}
\value{
A character string with the full path to the 'vep105_canonical.txt.gz' file.
}
\description{
This function returns the file path to the 'vep105_canonical.txt.gz' file included in the 'bravastring' package.
}
\examples{
canonical_file_path <- canonical_path()
print(canonical_file_path)
}
