% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Clean and Standardize Column Names}
\usage{
clean_string(column_names)
}
\arguments{
\item{column_names}{A character vector containing the column names to be cleaned.}
}
\value{
A character vector with the cleaned and standardized column names.
}
\description{
This function takes a vector of column names and returns a cleaned version of these names.
It standardizes the names by converting them to lowercase, replacing spaces, commas, and hyphens
with underscores, removing certain special characters, and replacing some symbols with text equivalents.
}
\examples{
column_names <- c("Column 1", "Column-2", "Column, 3", "Column&4")
clean_column_names <- clean_string(column_names)
print(clean_column_names)
# [1] "column_1" "column_2" "column_3" "column_or_4"
}
