% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_to_hgnc_symbol.R
\name{gene_id_to_hgnc_symbol}
\alias{gene_id_to_hgnc_symbol}
\title{Map Transcripts to Gene IDs to HGNC symbols}
\usage{
gene_id_to_hgnc_symbol(x)
}
\arguments{
\item{x}{A character vector of transcript IDs to be mapped to gene IDs.}
}
\value{
A character vector of hgnc IDs corresponding to the input ensembl gene IDs.
}
\description{
This function maps a given set of esnembl gene IDs to their corresponding hgnc_symbols.
}
