% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ancestries.R
\name{get_ancestries}
\alias{get_ancestries}
\title{Get Ancestry Encodings}
\usage{
get_ancestries()
}
\value{
A character vector containing the ancestry encodings: "eur", "eas", "sas", "afr", "jpn", "amr".
}
\description{
This function returns a character vector of common ancestry encodings.
}
\examples{
ancestries <- get_ancestries()
print(ancestries)
# [1] "eur" "eas" "sas" "afr" "jpn" "amr"
}
