% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biobanks.R
\name{get_biobanks}
\alias{get_biobanks}
\title{Get Biobank Names}
\usage{
get_biobanks()
}
\value{
A character vector containing the biobank names: "UKB", "GEL", "GNH", "BBJ", "PMB", and "BioMe".
}
\description{
This function returns a character vector of common biobank names.
}
\examples{
biobanks <- get_biobanks()
print(biobanks)
# [1] "UKB" "GEL" "GNH" "BBJ" "PMB" "BioMe"
}
