% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brava_description.R
\name{get_brava_description}
\alias{get_brava_description}
\title{Translate BRAVA Trait Identifier}
\usage{
get_brava_description(x)
}
\arguments{
\item{x}{A character vector of BRAVA phenotype identifiers to be translated.}
}
\value{
A character vector of translated BRAVA phenotype identifiers.
}
\description{
This function translates a BRAVA phenotype identifier to an alternative identifier format.
It reads the BRAVA phenotypes from a file and creates a mapping based on the identifiers.
}
