% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brava_samples.R
\name{get_brava_samples}
\alias{get_brava_samples}
\title{Get Brava Samples}
\usage{
get_brava_samples(biobank, ancestry, trait, sex = NULL, column = "N_eff")
}
\arguments{
\item{biobank}{Character. The name of the biobank.}

\item{ancestry}{Character. The ancestry of the samples.}

\item{trait}{Character. The trait of interest.}

\item{sex}{Character. The sex of the samples. Can be "F" (Female), "M" (Male), or "ALL" (default is NULL, which infers sex based on the trait).}

\item{column}{Character. The column to retrieve from the data (default is "N_eff").}
}
\value{
Numeric vector. The effective sample size for the specified parameters.
}
\description{
This function retrieves the effective sample size for a given biobank, ancestry, trait, and optionally sex.
}
\examples{
\dontrun{
get_brava_samples("ukb", "EUR", "Height")
get_brava_samples("ukb", "EUR", "Height", sex = "F")
}
}
