% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encodings.R
\name{get_encodings}
\alias{get_encodings}
\title{Get Genotype Encodings}
\usage{
get_encodings()
}
\value{
A character vector containing the genotype encodings: "recessive", "additive", and "dominance".
}
\description{
This function returns a character vector of common genotype encodings.
}
\examples{
encodings <- get_encodings()
print(encodings)
# [1] "recessive" "additive" "dominance"
}
