% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected_p.R
\name{get_expected_p}
\alias{get_expected_p}
\title{Get Expected P-values}
\usage{
get_expected_p(observed_p, na.rm = FALSE)
}
\arguments{
\item{observed_p}{A numeric vector of observed p-values.}

\item{na.rm}{A logical value indicating whether NA values should be removed. Default is FALSE.}
}
\value{
A numeric vector of expected p-values.
}
\description{
This function calculates the expected p-values based on observed p-values.
}
\examples{
observed_p <- c(0.01, 0.05, 0.03, NA)
get_expected_p(observed_p, na.rm = TRUE)
}
