% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd_chapters.R
\name{get_icd_chapters}
\alias{get_icd_chapters}
\title{Get ICD-10 Chapters and Code Ranges}
\usage{
get_icd_chapters()
}
\value{
A named list where each name is a chapter and each value is the corresponding ICD-10 code range.
}
\description{
This function returns a list of ICD-10 chapters along with their corresponding code ranges.
}
\examples{
icd_chapters <- get_icd_chapters()
print(icd_chapters)
}
