% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex_categories.R
\name{get_sex_categories}
\alias{get_sex_categories}
\title{Get Sex Categories}
\usage{
get_sex_categories()
}
\value{
A character vector containing the sex categories: "ALL", "FEMALE", "MALE", "M", "F", and "BOTH".
}
\description{
This function returns a character vector of common sex categories.
}
\examples{
sex_categories <- get_sex_categories()
print(sex_categories)
# [1] "ALL" "FEMALE" "MALE" "M" "F" "BOTH"
}
