% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex_specific_brava_traits.R
\name{get_sex_specific_brava_traits}
\alias{get_sex_specific_brava_traits}
\title{Get Sex-Specific Brava Traits}
\usage{
get_sex_specific_brava_traits(use_new_names = TRUE, single_sex = NULL)
}
\arguments{
\item{use_new_names}{Logical, whether to use new phenotype names. Defaults to TRUE.}

\item{single_sex}{Optional character, filter traits by a single sex ("Female" or "Male").}
}
\value{
A vector of phenotype IDs for traits that are not common to both sexes.
}
\description{
Retrieves phenotype IDs for sex-specific Brava traits, optionally filtered by a single sex.
}
