% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_symbol_path.R
\name{hgnc_symbol_path}
\alias{hgnc_symbol_path}
\title{Get Path to HGNC Ensembl Gene ID Map File}
\usage{
hgnc_symbol_path()
}
\value{
A character string with the full path to the 'ensgid_hgnc_symbol_chrom.txt.gz' file.
}
\description{
This function returns the file path to the 'ensgid_hgnc_symbol_chrom.txt.gz' file included in the 'bravastring' package.
}
