% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_sex_path.R
\name{icd_sex_path}
\alias{icd_sex_path}
\title{Get Path to BRAVA Phenotypes File}
\usage{
icd_sex_path()
}
\value{
A character string with the full path to the 'icd_sex_specific.txt' file.
}
\description{
This function returns the file path to the 'icd_sex_specific.txt' file included in the 'bravastring' package.
}
