% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_canonical.R
\name{is_canonical}
\alias{is_canonical}
\title{Check if Transcript is Canonical}
\usage{
is_canonical(x)
}
\arguments{
\item{x}{A character vector of transcript IDs to check for canonicity.}
}
\value{
A logical vector indicating whether each transcript ID is canonical.
}
\description{
This function checks if a given transcript is canonical based on the 'vep105_canonical.txt.gz' file.
}
\examples{
\dontrun{
transcript_ids <- c("ENST00000367770", "ENST00000494424")
canonical_status <- is_canonical(transcript_ids)
print(canonical_status)
}
}
