% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_icd_to_chapter.R
\name{map_icd_to_chapter}
\alias{map_icd_to_chapter}
\title{Map ICD-10 Codes to Chapters}
\usage{
map_icd_to_chapter(icd_code)
}
\arguments{
\item{icd_code}{A character string representing the ICD-10 code.}
}
\value{
A character string representing the chapter name corresponding to the given ICD-10 code.
}
\description{
This function maps two-digit ICD-10 codes to their corresponding chapters based on the predefined ranges.
}
\examples{
map_icd_to_chapter("A15") # Should return "Infectious"
map_icd_to_chapter("C34") # Should return "Neoplasms"
}
