% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_traits_path.R
\name{pilot_traits_path}
\alias{pilot_traits_path}
\title{Get Path to HGNC Ensembl Gene ID Map File}
\usage{
pilot_traits_path()
}
\value{
A character string with the full path to the 'brava_pilot_phenotypes.txt' file.
}
\description{
This function returns the file path to the 'brava_pilot_phenotypes.txt' file included in the 'bravastring' package.
}
