% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cauchy.R
\name{run_cauchy}
\alias{run_cauchy}
\title{Apply Cauchy Combination Test on Grouped Data}
\usage{
run_cauchy(
  grouped_dt,
  n_eff_name,
  Cauchy_stat_name,
  input_pvalues,
  output_meta_pvalue
)
}
\arguments{
\item{grouped_dt}{A grouped data table or data frame.}

\item{n_eff_name}{Character string. Name of the column containing effective sample sizes.}

\item{Cauchy_stat_name}{Character string. Name to be used for the output Cauchy statistic column.}

\item{input_pvalues}{Character string. Name of the column containing input p-values.}

\item{output_meta_pvalue}{Character string. Name to be used for the output meta-analysis p-value column.}
}
\value{
A data table with added columns:
  \item{Cauchy_stat_name}{The calculated Cauchy statistic.}
  \item{number_of_pvals}{The number of input p-values used.}
  \item{output_meta_pvalue}{The calculated meta-analysis p-value.}
}
\description{
This function performs a Cauchy combination test on grouped data, calculating
a meta-analysis p-value based on input p-values and sample sizes.
}
\details{
The function uses the Cauchy combination method to combine p-values, weighted
by the square root of the effective sample sizes. It handles extreme values 
and adjusts the final p-value to avoid values too close to 1.
}
