% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_heterogeneity_test.R
\name{run_heterogeneity_test}
\alias{run_heterogeneity_test}
\title{Perform Heterogeneity Test on Grouped Data}
\usage{
run_heterogeneity_test(grouped_dt, input_beta, output_meta_beta)
}
\arguments{
\item{grouped_dt}{A grouped data table or data frame. It should contain
columns for weights and the input beta coefficients.}

\item{input_beta}{Character string. Name of the column containing input beta coefficients.}

\item{output_meta_beta}{Character string. Name to be used for the output meta-analysis beta coefficient column.}
}
\value{
A data table with the following columns:
  \item{df}{Degrees of freedom (number of groups minus 1).}
  \item{sum_weights}{Sum of the weights.}
  \item{output_meta_beta}{The calculated meta-analysis beta coefficient.}
  \item{chisq_het}{Chi-square statistic for heterogeneity.}
  \item{Pvalue_het}{P-value for the heterogeneity test.}
}
\description{
This function conducts a heterogeneity test on grouped data, calculating
a meta-analysis beta coefficient and associated heterogeneity statistics.
}
\details{
The function calculates a weighted average of the input beta coefficients to produce
a meta-analysis beta. It then computes a chi-square statistic to test for 
heterogeneity among the input betas. The p-value for this test is calculated 
using the chi-square distribution with (n-1) degrees of freedom, where n is 
the number of groups.
}
