% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_ancestry.R
\name{str_extract_ancestry}
\alias{str_extract_ancestry}
\title{Extract Ancestry Annotation}
\usage{
str_extract_ancestry(x)
}
\arguments{
\item{x}{A character string from which to extract the ancestry annotation.}
}
\value{
A character string with the extracted ancestry annotation, or NA if no match is found.
}
\description{
This function extracts the ancestry annotation from a character string based on common ancestry encodings.
}
\examples{
extracted_ancestry <- str_extract_ancestry("example string containing eur ancestry")
print(extracted_ancestry)
# [1] "eur"
}
