% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_annotation.R,
%   R/str_extract_biobank.R
\name{str_extract_annotation}
\alias{str_extract_annotation}
\title{Extract Variant Annotation from a String}
\usage{
str_extract_annotation(x)

str_extract_annotation(x)
}
\arguments{
\item{x}{A character string from which to extract the biobank annotation.}
}
\value{
A character string with the extracted variant annotation, or NA if no match is found.

A character string with the extracted biobank, or NA if no match is found.
}
\description{
This function extracts a variant annotation from a given string using a regular expression pattern.
It constructs a regex pattern based on common variant annotations.

This function extracts a biobank annotation from a given string using a regular expression pattern.
It constructs a regex pattern based on common variant annotations.
}
\examples{
extracted_annotation <- str_extract_annotation("example string containing pLoF_damaging_missense")
print(extracted_annotation)
# [1] "pLoF_damaging_missense"
}
