% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_biobank.R
\name{str_extract_biobank}
\alias{str_extract_biobank}
\title{Extract Biobank from a String}
\usage{
str_extract_biobank(x)
}
\arguments{
\item{x}{A character string from which to extract the biobank annotation.}
}
\value{
A character string with the extracted biobank, or NA if no match is found.
}
\description{
This function extracts a biobank annotation from a given string using a regular expression pattern.
It constructs a regex pattern based on common variant annotations.
}
