% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_brava_trait.R
\name{str_extract_brava_trait}
\alias{str_extract_brava_trait}
\title{Extract BRAVA Trait from a String}
\usage{
str_extract_brava_trait(x, use_latest_names = TRUE)
}
\arguments{
\item{x}{A character string from which to extract the BRAVA phenotype identifier.}

\item{use_latest_names}{A logical value indicating whether to use the latest phenotype identifiers (default is TRUE).
If FALSE, uses an alternative set of identifiers.}
}
\value{
A character string with the extracted BRAVA phenotype identifier, or NA if no match is found.
}
\description{
This function extracts a BRAVA phenotype identifier from a given string using a regular expression pattern.
It reads the BRAVA phenotypes from a file and constructs a regex pattern based on the identifiers.
}
\examples{
extracted_trait <- str_extract_brava_trait("example string containing phenotype identifier")
print(extracted_trait)
}
