% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_brava_trait_gel.R
\name{str_extract_brava_trait_gel}
\alias{str_extract_brava_trait_gel}
\title{Extract BRAVA Trait from a String using GEL Biobank}
\usage{
str_extract_brava_trait_gel(x)
}
\arguments{
\item{x}{A character string from which to extract the BRAVA phenotype identifier.}
}
\value{
A character string with the extracted BRAVA phenotype identifier, or NA if no match is found.
}
\description{
This function extracts a BRAVA phenotype identifier from a given string using a regular expression pattern.
It reads the BRAVA phenotypes from a file and constructs a regex pattern based on the identifiers.
}
