% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_controls.R
\name{str_extract_controls}
\alias{str_extract_controls}
\title{Extract Number of Controls from a String}
\usage{
str_extract_controls(x)
}
\arguments{
\item{x}{A character string from which to extract the number of controls.}
}
\value{
A character string with the extracted number of controls, or NA if no match is found.
}
\description{
This function extracts the number of controls from a given string using a regular expression pattern.
It looks for a number that appears after a dot and is surrounded by either dots or underscores
in file naming patterns like "UKB.EUR.ALL.4490.218383.CervCanc".
}
\examples{
\dontrun{
extracted_controls <- str_extract_controls("UKB.EUR.ALL.4490.218383.CervCanc")
print(extracted_controls)
# [1] "218383"
}
}
