% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_encoding.R
\name{str_extract_encoding}
\alias{str_extract_encoding}
\title{Extract Genotype Encoding from a String}
\usage{
str_extract_encoding(x)
}
\arguments{
\item{x}{A character string from which to extract the genotype encoding.}
}
\value{
A character string with the extracted genotype encoding, or NA if no match is found.
}
\description{
This function extracts a genotype encoding from a given string using a regular expression pattern.
It constructs a regex pattern based on common genotype encodings.
}
\examples{
extracted_encoding <- str_extract_encoding("example string containing recessive")
print(extracted_encoding)
# [1] "recessive"
}
