% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_to_gene_id.R
\name{transcript_to_gene_id}
\alias{transcript_to_gene_id}
\title{Map Transcripts to Gene IDs}
\usage{
transcript_to_gene_id(x)
}
\arguments{
\item{x}{A character vector of transcript IDs to be mapped to gene IDs.}
}
\value{
A character vector of gene IDs corresponding to the input transcript IDs.
}
\description{
This function maps a given set of transcript IDs to their corresponding gene IDs using the 'vep105_canonical.txt.gz' file.
}
\examples{
\dontrun{
transcript_ids <- c("ENST00000367770", "ENST00000494424")
gene_ids <- transcript_to_gene_id(transcript_ids)
print(gene_ids)
}
}
