% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_brava_trait.R
\name{translate_brava_trait}
\alias{translate_brava_trait}
\title{Translate BRAVA Trait Identifier}
\usage{
translate_brava_trait(x)
}
\arguments{
\item{x}{A character vector of BRAVA phenotype identifiers to be translated.}
}
\value{
A character vector of translated BRAVA phenotype identifiers.
}
\description{
This function translates a BRAVA phenotype identifier to an alternative identifier format.
It reads the BRAVA phenotypes from a file and creates a mapping based on the identifiers.
}
